refactor: move ChatSocialMessage component
